♻️refactor to separate a function component into a individual file
