+++ lib/getStudyPost.jsx
@@ -1,0 +1,23 @@
+import path from "path";
+import fs from "fs";
+import matter from "gray-matter";
+
+const getStudyPost = async (slug) => {
+  const postPath = path.join(process.cwd(), "posts", "Study", `${slug}.mdx`);
+
+  if (!fs.existsSync(postPath)) {
+    return null; // 404(파일 없음))
+  }
+
+  const fileContent = fs.readFileSync(postPath, "utf-8");
+  const { content, data } = matter(fileContent);
+
+  return {
+    title: data.title,
+    description: data.description,
+    date: data.date,
+    content,
+  };
+};
+
+export default getStudyPost;
